Crate u_interface

Module containing different get request to the uHunt API

The requests are used to get certain data and are always organized into structs (or collections of a struct). This API request get data from the Universidad de Valladolid competitive programming judge.

This library can be used both in Rust and in python, as building it with maturin develop creates a python module.

Structs

Problem 🔒

A struct holding the data asociated with a problem

Submission 🔒

A Struct holding data for a problem submission

UserRank 🔒

A struct holding data for the position in the ranking of a user

UserSubmission 🔒

A struct holding data for an user’s submissions

Functions

PyInit_u_interface⚠

get_pdf_url_from_problem 🔒

Returns the URL of a pdf for a certain problem number.

get_problem 🔒

Returns the problem with number num, or an empty struct if it does not exist. This function is async, that means it has to be called inside an async function/method or using tokio::Runtime.block_on().

get_problem_by_pid 🔒

Returns the problem with id pid, or an empty struct if it does not exist. This function is async, that means it has to be called inside an async function/method or using tokio::Runtime.block_on().

get_ranking 🔒

Returns a Vector containing the Rank of the above users over the user with id uid, the user with that id and the below users, below the asme user. This function is async, that means it has to be called inside an async function/method or using tokio::Runtime.block_on().

get_submissions_problem 🔒

Returns a vector containing the submissions to a certain problem in a certain timeframe. This function is async, that means it has to be called inside an async function/method or using tokio::Runtime.block_on().

get_uid_from_uname 🔒

Returns the user’s id number by searching using the stated username. This function is async, that means it has to be called inside an async function/method or using tokio::Runtime.block_on().

get_user_submissions 🔒

Returns the user’s last submissions to any problem. This function is async, that means it has to be called inside an async function/method or using tokio::Runtime.block_on().

get_usubmissions_to_problem 🔒

Returns a HashMap containing the submissions the speccified user to the selected problem. This function is async, that means it has to be called inside an async function/method or using tokio::Runtime.block_on().